+++ tests/2xx/253_csl_get_version.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+. common.sh
+
+assert_func csl_get_version "${TEST_OK}" "${CSL_VERSION}"